refactor : signin refactoring &  add mail file
